tests/models/krate: Move test code into `index` module
