+++ src/tests/mod.rs
@@ -14,14 +14,13 @@ use diesel::prelude::*;
 mod account_lock;
 mod authentication;
 mod blocked_routes;
-mod builders;
+pub mod builders;
 mod categories;
 mod cors;
 mod dump_db;
 mod github_secret_scanning;
 mod krate;
 mod middleware;
-mod models;
 mod not_found_error;
 mod owners;
 mod pagination;
