--- src/tests/models/mod.rs
@@ -1,1 +1,0 @@
-mod krate;
